requirements.txt
absl_py==1.4.0
Flask==2.3.2
Flask_RESTful==0.3.9
marshmallow==3.19.0
six==1.16.0
